added character list view
The CharacterListViewViewModel file contains the data processing and display logic associated with the character list view. A function was defined to fetch data from the API and it conforms to the UICollectionViewDataSource, UICollectionViewDelegate, and UICollectionViewDelegateFlowLayout protocols. In the CharacterListView file, the creation and configuration of the character list view were defined. The spinner view is used to indicate loading status to the user while character data is being fetched. Finally, in the RMCharacterViewController file, the addition of the character list view and its placement in the four corners of the screen was ensured through the setUpView() function.