added character cell
The RMCharacterCollectionViewCell file  represents a single cell for displaying a character within a collection view. The cell contains an UIImageView for displaying the character's image, and two labels for the character's name and status. The RMCharacterCollectionViewCellViewModel file manages the view models for cells displaying characters. It contains information such as the character's name, status, and URL for the image.