characters displayed
The RMCharacterListViewViewModelDelegate defines a protocol for objects that need to be notified when the initial character data is loaded. Conforming objects implement the didLoadInitialCharacters() method to handle this event. This method is called upon completion of the initial character loading process, updating visual elements to indicate the loading is finished.