character photo cell
This cell includes a UIImageView with specific properties such as contentMode, clipsToBounds, and cornerRadius set up for displaying character images. The cell manages image fetching through a view model that leverages RMImageLoader for downloading images asynchronously. Constraints for the imageView are set up to anchor it to the edges of the contentView. Additionally, the prepareForReuse method is overridden to reset the imageView's image when the cell is being reused.