character info view model
RMCharacterInfoCollectionViewCellViewModel utilizes an enum named Type, which determines the type of character attribute and provides appropriate icons, colors, and titles for various attributes such as status, gender, type, etc. Different icons and colors are defined for these characteristics. Additionally, two DateFormatter instances named dateFormatter and shortDateFormatter have been added, which are used to present date and time information in appropriate formats. The displayValue method converts the value into a readable format if it is a date; otherwise, it returns the value directly. This configuration aims to centrally and orderly manage the presentation and processing of each character attribute.