added episode detail view controller
Modifications have been made to the collectionView(_:didSelectItemAt:) method to handle episode selection based on the section type and navigate to the detail view controller. Additionally, the RMEpisodeDetailViewController class has been created to display details of a selected episode, initializing with the episode's URL and setting up the user interface accordingly. These updates aim to enhance the user experience within the application.