api cache manager
The manager utilizes a dictionary to store NSCache instances associated with each RMEndpoint. It includes methods to retrieve cached data (cachedResponse(for:url:)) and to store data (setCache(for:url:data:)) in the respective cache based on the endpoint. This caching mechanism aims to enhance performance by reducing the frequency of network requests and speeding up data retrieval in the app.